refactor payments module to mvc
