+++ backend/payments/routes.php
@@ -1,10 +1,12 @@
 <?php
 require_once(__DIR__.'/../vendor/autoload.php');
-include __DIR__.'/index.php';
+session_start();
 
 use Vendor\Schoolarsystem\DBConnection;
+use Vendor\Schoolarsystem\Controllers\PaymentsController;
 
 $connection = new DBConnection();
+$payments = new PaymentsController($connection);
 
 if($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['action'])){
 
@@ -13,63 +15,42 @@ if($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['action'])){
     switch ($action) {
         case 'AddPayment':
             $paymentData = $_POST['data'];
-
             parse_str($paymentData, $paymentDataArray);
-
-            $addPayment = new PaymentsControl($connection);
-            $add = $addPayment->AddPayment($paymentDataArray);
-
+            $add = $payments->addPayment($paymentDataArray);
             header('Content-Type: application/json');
             echo json_encode($add);
             break;
-        
+
         case 'VerifyTaxData':
             $studentId = $_POST['data'];
-
-            $verifyTaxData = new PaymentsControl($connection);
-            $verify = $verifyTaxData->VerifyTaxData($studentId);
-
+            $verify = $payments->verifyTaxData($studentId);
             header('Content-Type: application/json');
             echo json_encode($verify);
             break;
 
         case 'getStudentsPayMount':
-
-            $getStudentsPayMount = new PaymentsControl($connection);
-            $get = $getStudentsPayMount->GetStudentsPayMount();
-
+            $get = $payments->getStudentsPayMount();
             header('Content-Type: application/json');
             echo json_encode($get);
             break;
 
         case 'setStudentPayMount':
-
             $data = $_POST['data'];
             $studentId = $data['studentId'];
             $amount = $data['amount'];
-
-            
-            $setStudentPayMount = new PaymentsControl($connection);
-            $set = $setStudentPayMount->SetStudentPayMount($studentId, $amount);
-            
+            $set = $payments->setStudentPayMount($studentId, $amount);
             header('Content-Type: application/json');
             echo json_encode($set);
             break;
 
         case 'VerifyMonthlyPayment':
             $studentId = $_POST['data'];
-
-            $verifyMonthlyPayment = new PaymentsControl($connection);
-            $verify = $verifyMonthlyPayment->VerifyMonthlyPayment($studentId);
-
+            $verify = $payments->verifyMonthlyPayment($studentId);
             header('Content-Type: application/json');
             echo json_encode($verify);
             break;
 
         default:
-            # code...
             break;
     }
-
 }
-?>
